requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
install==1.3.5
Pillow==10.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
